fix: padStart/padEnd length coercion — ToLength, not FPToSI(NaN) UB
'abc'.padEnd(NaN, 'def') must give 'abc' (ToLength(ToIntegerOrInfinity(NaN))
= 0), but the maxLength reached ts_string_pad{Start,End}'s int64_t param via
a bare FPToSI (compiled path) / ts_value_get_int (native path). FPToSI on
NaN is undefined; depending on prior FP register state it yielded a garbage
huge positive length -> a multi-GB allocation -> OOM (or a multi-second
hang). That context-dependence is why the padEnd/padStart
max-length-not-greater-than-string tests oscillated pass/timeout/fail in
essentially every sweep and were mis-labeled "perf flakes."

Pass the raw double through to the runtime (lowering: i64Arg ->
f64Arg(ToF64); native wrapper: ts_value_get_double) and apply
ToLength(ToIntegerOrInfinity) once, centrally, in ts_string_pad{Start,End}:
NaN/<=0 -> 0, fractional -> truncate, +Inf/>=2^53-1 -> 2^53-1.

Gates: golden-ir 267/279, node 295/297, sample 1831/236. Full sweep
26,935 -> 26,938 (+3: padEnd + padStart max-length now pass
deterministically, and replace-math — another flake-family victim —
stabilizes; 0 regressions). Removes the recurring phantom downs from
future sweeps.

+++ src/runtime/include/TsString.h
@@ -121,8 +121,8 @@ extern "C" {
     void* ts_string_substring(void* str, int64_t start, int64_t end);
     void* ts_string_slice(void* str, int64_t start, int64_t end);
     void* ts_string_repeat(void* str, int64_t count);
-    void* ts_string_padStart(void* str, int64_t targetLength, void* padString);
-    void* ts_string_padEnd(void* str, int64_t targetLength, void* padString);
+    void* ts_string_padStart(void* str, double targetLength, void* padString);
+    void* ts_string_padEnd(void* str, double targetLength, void* padString);
     bool ts_string_startsWith(void* str, void* prefix);
     bool ts_string_endsWith(void* str, void* suffix);
     bool ts_string_includes(void* str, void* searchString);
